Maxl/pyproject.toml
[project]
name = "maxl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "mysqlclient>=2.2.8",
]
